Re-write extender commands to be subclasses; add RaiseToClimb
