+++ src/main/java/org/usfirst/frc/team4373/robot/commands/RaiseToClimb.java
@@ -1,0 +1,11 @@
+package org.usfirst.frc.team4373.robot.commands;
+
+public class RaiseToClimb extends VerticalExtenderSetter {
+
+    /**
+     * Constructs a new RaiseToClimb command.
+     */
+    protected RaiseToClimb() {
+        super("RaiseToClimb", 33, 0);
+    }
+}
